src/main/java/com/nathan/app/GreetingController.java
package com.nathan.app;

import java.util.concurrent.atomic.AtomicLong;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class GreetingController {
	private static final String template = "Hello, %s!";
	private final AtomicLong counter = new AtomicLong();
	
	// request matches method name
	// requestparam name is equal to specified, or else default
	@RequestMapping(value="/greeting", method=RequestMethod.GET, produces="application/json")
	public Greeting greeting (@RequestParam(value="name", defaultValue="World") String name) {
		return new Greeting(counter.incrementAndGet(),
			String.format(template, name));
	}
}